requirements.txt
Django==4.2.16
djangorestframework==3.15.2
Pillow==10.3.0
python-decouple==3.8
razorpay==1.4.1
reportlab==4.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
requests==2.31.0
geopy==2.4.1
pytesseract==0.3.10
gunicorn==21.2.0
